subida y descarga de archivos

+++ src/app/pages/mantenimientos/medicos/medicos.component.ts
@@ -81,22 +81,22 @@ export class MedicosComponent implements OnInit ,OnDestroy{
 
  eliminarMedico( medico:Medico){
 
-//dialogo predefinido en swetAlert2 y adaptado a la aplicacion
-Swal.fire({
-  title: '¿Borrar Medico?',
-  text: `Esta a punto de eliminar a ${ medico.nombre}`,
-  icon: 'question',
-  showCancelButton: true,
-  confirmButtonColor: '#3085d6',
-  cancelButtonColor: '#d33',
-  confirmButtonText: 'Si, borrar'
-}).then((result) => {
-  if (result.isConfirmed) {
-    this.medicosService.eliminarMedico(medico._id).subscribe(
-      resp=>{
-        this.cargarMedicos();
-        Swal.fire('Borrado', medico.nombre,'success')
-      }
+  //dialogo predefinido en swetAlert2 y adaptado a la aplicacion
+    Swal.fire({
+       title: '¿Borrar Medico?',
+       text: `Esta a punto de eliminar a ${ medico.nombre}`,
+       icon: 'question',
+       showCancelButton: true,
+       confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, borrar'
+     }).then((result) => {
+      if (result.isConfirmed) {
+        this.medicosService.eliminarMedico(medico._id).subscribe(
+          resp=>{
+           this.cargarMedicos();
+           Swal.fire('Borrado', medico.nombre,'success')
+       }
     )
       Swal.fire(
         'Borrado!',
